submission_in_python/pyproject.toml
[virtualenvs]
prefer-active-python = true

[tool.poetry]
name = "explore-the-blockchain-python"
version = "0.0.1"
description = "A fork of bitcoin-dev-launchpad-2024-explore-the-blockchain-task-rpc-scavenger-hunt created by GitHub Classroom in python"
authors = ["qlrd"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-bitcoinrpc = "^1.0"
python-bitcoinlib = "^0.12.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.1"
poethepoet = "^0.30.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
format = "black ./src"
lint = "pylint ./src"
test = "pytest ./tests"
submission_001 = "python ./src/submission_001.py"
submission_002 = "python ./src/submission_002.py"
submission_003 = "python ./src/submission_003.py"
submission_004 = "python ./src/submission_004.py"
submission_005 = "python ./src/submission_005.py"
submission_006 = "python ./src/submission_006.py"
submission_007 = "python ./src/submission_007.py"
submission_008 = "python ./src/submission_008.py"
